# test: zero-copy tranformation from struct to bytes