@gangliao 2021-03-18
## Old solution:

```bash
Raw Arrow Flight data: 3453248, encoding ratio: 1.350
Compression time: 89 ms
Compressed Arrow Flight data: 1532739, type: Snappy, compression ratio: 3.041
Decompression time: 36 ms
Compression time: 5 ms
Compressed Arrow Flight data: 1552275, type: Lz4, compression ratio: 3.003
Decompression time: 1 ms
Compression time: 16 ms
Compressed Arrow Flight data: 944340, type: Zstd, compression ratio: 4.936
Decompression time: 4 ms
```

## New solution

```bash
transmute data - raw data: 3453248
transmute data - compression time: 430 us
transmute data - compressed data: 1772, type: Zstd
transmute data - decompression time: 261 us
transmute data - decompress raw data: 3453248
```

compression ratio: **1948.78555305** !!!
Serialization (including compression and decompression) is also an order of magnitude faster